filter-repo: add --analyze option
This option walks through the repository history and creates a report
with basic statistics, rename related information, and sizes of objects
and when/if those have been deleted.  It primarily looks at unpacked
sizes (i.e. size of object ignoring delta-ing and compression), and
sums the size of each version of the file for each path.  Additionally,
it aggregates these sums by extension and by directory, and tracks
whether paths, extensions, and directories have been deleted.  This can
be very useful in determining what the big things are, and whether they
might have been considered to have been mistakes to add to the
repository in the first place.

There are numerous caveats with the determination of "deleted" and
"renamed", and can give both false positives and false negatives.  But
they are only meant as a helpful heuristic to give others a starting
point for an investigation, and the information provide so far is useful.
I do want to improve the equivalence classes (rename handling), but that
is for a future commit.

Signed-off-by: Elijah Newren <newren@gmail.com>