filter-repo: use exported/imported refs for cleanup and metadata recording
Now that we are tracking exported and imported refs, we no longer need
to rely on _orig_refs and _seen_refs for deletion of "unused" refs at
the end of the run.  Verify that we correctly tracked exported and
imported refs by using them instead for the post-run ref deletion.  This
removes the last use of _seen_refs, which will be removed in a
subsequent commit.

Signed-off-by: Elijah Newren <newren@gmail.com>