filter-repo: make original changes available to callbacks
Scripts making use of filter-repo as a library may want to be able to
compare the list of file changes in a commit to the original list of
file changes before other filtering occurred.  For example, an
lfs-conversion script might want to know that e.g. --path selection and
--filename-callback said a bunch of files should be removed, so that the
lfs-conversion script can instead load those files up to git lfs.

FIXME: This slows down rails rewriting by over 60%, which is unacceptable;
need to find a different way to introduce this change.  Maybe only when
explicitly requested somehow?

Signed-off-by: Elijah Newren <newren@gmail.com>